+++ dallinger/data.py
@@ -13,6 +13,7 @@ from zipfile import ZipFile, ZIP_DEFLATED
 import boto
 from boto.s3.key import Key
 import hashlib
+import postgres_copy
 import psycopg2
 
 try:
@@ -24,6 +25,7 @@ except ImportError:
 
 from dallinger import heroku
 from dallinger import db
+from dallinger import models
 
 
 table_names = [
@@ -39,9 +41,34 @@ table_names = [
 ]
 
 
-def load(id):
-    """Load the data from wherever it is found."""
-    data_filename = '{}-data.zip'.format(id)
+def find_experiment_export(app_id):
+    """Attempt to find a zipped export of an experiment with the ID provided
+    and return its path. Returns None if not found.
+
+    Search order:
+        1. local "data" subdirectory
+        2. user S3 bucket
+        3. Dallinger S3 bucket
+    """
+
+    # Check locally first
+    cwd = os.getcwd()
+    data_filename = '{}-data.zip'.format(app_id)
+    path_to_data = os.path.join(cwd, "data", data_filename)
+    if os.path.exists(path_to_data):
+        try:
+            Data(path_to_data)
+        except IOError:
+            from dallinger import logger
+            logger.exception(
+                "Error reading local data file {}, checking remote.".format(
+                    path_to_data
+                )
+            )
+        else:
+            return path_to_data
+
+    # Get remote file instead
     path_to_data = os.path.join(tempfile.mkdtemp(), data_filename)
 
     buckets = [
@@ -54,11 +81,19 @@ def load(id):
         k.key = data_filename
         try:
             k.get_contents_to_filename(path_to_data)
-            return Data(path_to_data)
         except boto.exception.S3ResponseError:
             pass
+        else:
+            return path_to_data
 
-    raise IOError("Dataset {} could not be found.".format(id))
+
+def load(app_id):
+    """Load the data from wherever it is found."""
+    path_to_data = find_experiment_export(app_id)
+    if path_to_data is None:
+        raise IOError("Dataset {} could not be found.".format(app_id))
+
+    return Data(path_to_data)
 
 
 def dump_database(id):
@@ -101,6 +136,27 @@ def backup(id):
     return url
 
 
+def registration_key(id):
+    return '{}.reg'.format(id)
+
+
+def register(id, url=None):
+    """Register a UUID key in the global S3 bucket."""
+    k = Key(registration_s3_bucket())
+    k.key = registration_key(id)
+    k.set_contents_from_string(url or 'missing')
+    reg_url = k.generate_url(expires_in=0, query_auth=False)
+    return reg_url
+
+
+def is_registered(id):
+    """Check if a UUID is already registered"""
+    # We can't use key.exists() unless the user has GET access,
+    # exists() would scale much better though.
+    key_names = set(k.key for k in list(registration_s3_bucket().list()))
+    return registration_key(id) in key_names
+
+
 def copy_heroku_to_local(id):
     """Copy a Heroku database locally."""
 
@@ -133,6 +189,7 @@ def copy_local_to_csv(local_db, path, scrub_pii=False):
         with open(csv_path, "w") as f:
             sql = "COPY {} TO STDOUT WITH CSV HEADER".format(table)
             cur.copy_expert(sql, f)
+
     if scrub_pii:
         _scrub_participant_table(path)
 
@@ -210,8 +267,56 @@ def export(id, local=False, scrub_pii=False):
         k = Key(user_s3_bucket())
         k.key = data_filename
         k.set_contents_from_filename(path_to_data)
+        url = k.generate_url(expires_in=0, query_auth=False)
+
+        # Register experiment UUID with dallinger
+        register(id, url)
 
     return path_to_data
+
+
+def ingest_zip(path):
+    """Given a path to a zip file created with `export()`, recreate the
+    database with the data stored in the included .csv files.
+    """
+    import_order = [
+        "network",
+        "participant",
+        "node",
+        "info",
+        "notification",
+        "question",
+        "transformation",
+        "vector",
+        "transmission",
+    ]
+    with ZipFile(path, 'r') as archive:
+        filenames = archive.namelist()
+        for name in import_order:
+            filename = [f for f in filenames if name in f][0]
+            model_name = name.capitalize()
+            model = getattr(models, model_name)
+            file = archive.open(filename)
+            ingest_to_model(file, model)
+
+
+def fix_autoincrement(table_name):
+    """Auto-increment pointers are not updated when IDs are set explicitly,
+    so we manually update the pointer so subsequent inserts work correctly.
+    """
+    db.engine.execute(
+        "select setval('{0}_id_seq', max(id)) from {0}".format(table_name)
+    )
+
+
+def ingest_to_model(file, model):
+    """Load data from a CSV file handle into storage for a
+    SQLAlchemy model class.
+    """
+    postgres_copy.copy_from(
+        file, model, db.engine, format='csv', HEADER=True
+    )
+    fix_autoincrement(model.__table__.name)
 
 
 def archive_data(id, src, dst):
@@ -235,10 +340,17 @@ def user_s3_bucket(canonical_user_id=None):
     s3_bucket_name = "dallinger-{}".format(
         hashlib.sha256(canonical_user_id).hexdigest()[0:8])
 
+    config = get_config()
+    location = config.get('aws_region')
+
+    # us-east-1 is the default and should not be included as location
+    if not location or location == u'us-east-1':
+        location = boto.s3.connection.Location.DEFAULT
+
     if not conn.lookup(s3_bucket_name):
         bucket = conn.create_bucket(
             s3_bucket_name,
-            location=boto.s3.connection.Location.DEFAULT
+            location=location
         )
     else:
         bucket = conn.get_bucket(s3_bucket_name)
@@ -248,18 +360,25 @@ def user_s3_bucket(canonical_user_id=None):
 
 def dallinger_s3_bucket():
     """The public `dallinger` S3 bucket."""
-    conn = _s3_connection()
+    conn = _s3_connection(dallinger_region=True)
     return conn.get_bucket("dallinger")
 
 
-def _s3_connection():
+def registration_s3_bucket():
+    """The public write-only `dallinger-registration` S3 bucket."""
+    conn = _s3_connection(dallinger_region=True)
+    return conn.get_bucket("dallinger-registrations")
+
+
+def _s3_connection(dallinger_region=False):
     """An S3 connection using the AWS keys in the config."""
     config = get_config()
     if not config.ready:
         config.load()
 
+    region = 'us-east-1' if dallinger_region else config.get('aws_region')
     return boto.s3.connect_to_region(
-        config.get('aws_region'),
+        region,
         aws_access_key_id=config.get('aws_access_key_id'),
         aws_secret_access_key=config.get('aws_secret_access_key'),
     )