+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Change Log
 
-## [v3.1.0](https://github.com/dallinger/dallinger/tree/v3.0.1) (2017-06-23)
+## [v3.2.0](https://github.com/dallinger/dallinger/tree/v3.2.0) (2017-06-25)
 
-- **BREAKING** The fitness parameter of `dallinger.nodes.Agent` is now a
-  floating point number, rather than an integer.
+- **Feature** Datasets in zip files can now be import to an experiment server
+  via a new CLI command, `dallinger load`.
+
+## [v3.1.0](https://github.com/dallinger/dallinger/tree/v3.1.0) (2017-06-23)
 
 - **FEATURE** The `dallinger uuid` command line tool has been added to generate
   a UUID for an experiment.
@@ -57,6 +59,9 @@
 
 - A bug in the chatroom demo was fixed which improves 
   reliability and simplifies the connection to the backend. [\#537](https://github.com/Dallinger/Dallinger/issues/537)
+
+- The fitness parameter of `dallinger.nodes.Agent` is now a
+  floating point number, rather than an integer.
 
 ## [v3.0.1](https://github.com/dallinger/dallinger/tree/v3.0.1) (2017-06-19)